+++ src/main/java/io/github/albertsongs/videoreceiversmanager/controller/v1/VideoControllerV1.java
@@ -9,16 +9,15 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 @RequestMapping("api/v1/videos")
 @CrossOrigin("https://albertsongs.github.io")
-public final class VideoController {
+public final class VideoControllerV1 {
     @Autowired
     private VideoService videoService;
 
     @GetMapping
-    public ObjectListContainer<Video> getAllVideos(@RequestParam(name = "playlistId", required = false) Long playlistId) {
-        final ObjectListContainer<Video> videos = new ObjectListContainer<>();
-        videos.setList(playlistId == null
+    public ObjectListContainer<Video> getAllVideos(
+            @RequestParam(name = "playlistId", required = false) Long playlistId) {
+        return new ObjectListContainer<>(playlistId == null
                 ? videoService.getAll()
                 : videoService.getAllFromPlaylistById(playlistId));
-        return videos;
     }
 }
