+++ src/main/java/io/github/albertsongs/videoreceiversmanager/model/ObjectListContainer.java
@@ -1,8 +1,10 @@
 package io.github.albertsongs.videoreceiversmanager.model;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
 
 @Data
+@AllArgsConstructor
 public class ObjectListContainer<Object> {
     private Iterable<Object> list;
 }
