+++ src/main/java/io/github/albertsongs/videoreceiversmanager/model/VideoEx.java
@@ -8,6 +8,17 @@ public final class VideoEx extends Video {
     @Getter
     @Setter
     private VideoMeta youtube;
+
+    public VideoEx(Long id, String title, String url, String subtitlesUrl, VideoMeta youtube) {
+        super(id, title, url, subtitlesUrl);
+        this.youtube = youtube;
+    }
+
+    public VideoEx(VideoEntity entity, VideoMeta youtube) {
+        super(entity);
+        this.youtube = youtube;
+    }
+
     public VideoEx(VideoEntity entity) {
         super(entity);
         youtube = new VideoMeta(entity.getYoutubeId(), entity.getPlaylist().getYoutubeId());
